fix: harden app_logic/can_formatter APIs against misuse
- app_can_message.timestamp_ms is now uint64_t (k_uptime_get()'s type);
  uint32_t would wrap around after ~49.7 days
- can_formatter_format() rejects dlc > APP_CAN_MESSAGE_MAX_DLC with
  CAN_FORMATTER_ERR_INVALID_DLC instead of silently truncating the
  payload while still printing the original, inconsistent DLC value
- can_formatter_format() now returns an explicit
  enum can_formatter_result (ok / null arg / invalid DLC / buffer too
  small) plus an out_len, instead of overloading a bare snprintf-style
  int for every failure mode
- app_logic_init() returns false on a NULL argument instead of
  dereferencing it; app_logic_handle_message() returns the new
  APP_LOGIC_ACTION_INVALID_ARGS instead of dereferencing state/message

+++ app/include/specialized/app_can_message.h
@@ -15,7 +15,7 @@ struct app_can_message {
 	bool extended_id;
 	uint8_t dlc;
 	uint8_t data[APP_CAN_MESSAGE_MAX_DLC];
-	uint32_t timestamp_ms;
+	uint64_t timestamp_ms;
 };
 
 #endif /* SPECIALIZED_APP_CAN_MESSAGE_H_ */
